ref: split audio engine external fdk cluster
